feat: add pylint and autopep-8
