+++ .pylintrc
@@ -1,0 +1,8 @@
+[MESSAGES CONTROL]
+  disable=
+    C0114, # missing-module-docstring
+    C0115, # missing-class-docstring
+    C0116 # missing-function-docstring
+
+[MASTER]
+  init-hook='import sys; sys.path.append("./app")'
