+++ app/core.py
@@ -1,0 +1,10 @@
+from dotenv import load_dotenv
+
+from main.application import Application
+
+if __name__ == "__main__":
+    load_dotenv()
+
+    app = Application()
+
+    app.run()
