--- module/config/dotenv.py
@@ -1,3 +1,0 @@
-from dotenv import load_dotenv
-
-load_dotenv()
