--- module/core.py
@@ -1,6 +1,0 @@
-from main.application import Application
-import config.dotenv
-
-app = Application()
-
-app.run()
