+++ requirements.txt
@@ -1,12 +1,26 @@
 aiohttp==3.8.3
 aiosignal==1.2.0
+astroid==2.12.10
 async-timeout==4.0.2
 attrs==22.1.0
+autopep8==1.7.0
 charset-normalizer==2.1.1
+dill==0.3.5.1
 discord.py==2.0.1
 frozenlist==1.3.1
 idna==3.4
+isort==5.10.1
+lazy-object-proxy==1.7.1
+mccabe==0.7.0
 multidict==6.0.2
 pkg_resources==0.0.0
+platformdirs==2.5.2
+pycodestyle==2.9.1
+pylint==2.15.3
 python-dotenv==0.21.0
+toml==0.10.2
+tomli==2.0.1
+tomlkit==0.11.5
+typing_extensions==4.3.0
+wrapt==1.14.1
 yarl==1.8.1
